Initial attempts at the program during HackRU 2017
Due to errors in the laptop, i was unable to proceed, or even fully check, if the initial attempt was successful. A lot of errors in my part to not fully update a laptop that hasn`t been touched in a year. Regardless, i plan to transfer the data over to my home station, search for a better API, and learn to utilize MongoDB, like i had initially planned to.